[ROX-5805] Compute Node Risk (stackrox/rox#7166)

+++ central/risk/multipliers/utils.go
@@ -3,7 +3,7 @@ package multipliers
 // NormalizeScore normalizes score into a value between 1 and maxAllowedValue.
 func NormalizeScore(score, saturation, maxAllowedValue float32) float32 {
 	if score > saturation {
-		return float32(maxAllowedValue)
+		return maxAllowedValue
 	}
 	return 1 + (score/saturation)*(maxAllowedValue-1)
 }
